+++ docs/source/data_collections.rst
@@ -1,0 +1,6 @@
+data_collections
+================
+
+.. automodule:: sentinelhub.data_collections
+    :members:
+    :show-inheritance:
