--- docs/source/data_sources.rst
@@ -1,6 +1,0 @@
-data_sources
-============
-
-.. automodule:: sentinelhub.data_sources
-    :members:
-    :show-inheritance:
