+++ tests/test_ogc.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from shapely.geometry import MultiPolygon
 
-from sentinelhub import WmsRequest, WcsRequest, CRS, MimeType, CustomUrlParam, ServiceType, DataSource, BBox,\
+from sentinelhub import WmsRequest, WcsRequest, CRS, MimeType, CustomUrlParam, ServiceType, DataCollection, BBox,\
     WebFeatureService, DownloadFailedException, TestSentinelHub, TestCaseContainer
 from sentinelhub.data_request import OgcRequest
 from sentinelhub.ogc import OgcImageService
@@ -109,41 +109,41 @@ class TestOgc(TestSentinelHub):
                                                           CustomUrlParam.GEOMETRY: geometry_wkt_pop_web}),
                             result_len=1, img_min=64, img_max=MimeType.PNG.get_expected_max_value(), img_mean=213.6124,
                             img_median=242.0, tile_num=3),
-            # DataSource tests:
+            # DataCollection tests:
             cls.OgcTestCase('S2 L1C Test',
-                            WmsRequest(data_source=DataSource.SENTINEL2_L1C, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.SENTINEL2_L1C, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d8, layer='BANDS-S2-L1C',
                                        width=img_width, height=img_height, bbox=wgs84_bbox,
                                        time=('2017-10-01', '2017-10-02')),
                             result_len=1, img_min=0, img_max=160, img_mean=60.50514, img_median=63.0,
                             tile_num=2),
             cls.OgcTestCase('S2 L2A Test',
-                            WmsRequest(data_source=DataSource.SENTINEL2_L2A, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.SENTINEL2_L2A, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF, layer='BANDS-S2-L2A',
                                        width=img_width, height=img_height, bbox=wgs84_bbox,
                                        time=('2017-10-01', '2017-10-02')),
                             result_len=1, img_min=0.0, img_max=MimeType.TIFF.get_expected_max_value(),
                             img_mean=22744.01498, img_median=21391.0, tile_num=2),
             cls.OgcTestCase('L8 Test',
-                            WmsRequest(data_source=DataSource.LANDSAT8, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.LANDSAT8, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='BANDS-L8',
                                        width=img_width, height=img_height, bbox=wgs84_bbox,
                                        time=('2017-10-05', '2017-10-10'), time_difference=datetime.timedelta(hours=1)),
                             result_len=1, img_min=0.0012, img_max=285.78003, img_mean=52.06088, img_median=0.5192,
                             tile_num=2),
             cls.OgcTestCase('DEM Test',
-                            WmsRequest(data_source=DataSource.DEM, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.DEM, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='DEM',
                                        width=img_width, height=img_height, bbox=wgs84_bbox),
                             result_len=1, img_min=-108.0, img_max=-18.0, img_mean=-72.2097, img_median=-72.0),
             cls.OgcTestCase('MODIS Test',
-                            WmsRequest(data_source=DataSource.MODIS, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.MODIS, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='BANDS-MODIS',
                                        width=img_width, height=img_height, bbox=wgs84_bbox, time='2017-10-01'),
                             result_len=1, img_min=0.0, img_max=3.2767, img_mean=0.136548, img_median=0.00240,
                             tile_num=1),
             cls.OgcTestCase('S1 IW Test',
-                            WmsRequest(data_source=DataSource.SENTINEL1_IW, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.SENTINEL1_IW, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='BANDS-S1-IW',
                                        width=img_width, height=img_height, bbox=wgs84_bbox,
                                        time=('2017-10-01', '2017-10-02'),
@@ -151,14 +151,14 @@ class TestOgc(TestSentinelHub):
                             result_len=1, img_min=0.00150, img_max=MimeType.TIFF_d32f.get_expected_max_value(),
                             img_mean=0.1135815, img_median=0.06210, tile_num=2),
             cls.OgcTestCase('S1 EW Test',
-                            WmsRequest(data_source=DataSource.SENTINEL1_EW, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.SENTINEL1_EW, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='BANDS-S1-EW',
                                        width=img_width, height=img_height, bbox=wgs84_bbox_2,
                                        time=('2018-2-7', '2018-2-8'),
                                        time_difference=datetime.timedelta(hours=1)),
                             result_len=2, img_min=0.0003, img_max=1.0, img_mean=0.53118, img_median=1.0, tile_num=3),
             cls.OgcTestCase('S1 EW SH Test',
-                            WmsRequest(data_source=DataSource.SENTINEL1_EW_SH,
+                            WmsRequest(data_collection=DataCollection.SENTINEL1_EW_SH,
                                        data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d16, layer='BANDS-S1-EW-SH',
                                        width=img_width, height=img_height, bbox=wgs84_bbox_3,
@@ -167,7 +167,7 @@ class TestOgc(TestSentinelHub):
                             result_len=1, img_min=472, img_max=3538, img_mean=926.8818, img_median=894.0,
                             tile_num=1),
             cls.OgcTestCase('S1 EW ASC Test',
-                            WmsRequest(data_source=DataSource.SENTINEL1_EW_ASC, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.SENTINEL1_EW_ASC, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='BANDS-S1-EW',
                                        width=img_width, height=img_height, bbox=wgs84_bbox_2,
                                        time=('2018-2-7', '2018-2-8'),
@@ -175,7 +175,7 @@ class TestOgc(TestSentinelHub):
                             result_len=1, img_min=0.0003, img_max=0.2322, img_mean=0.02199, img_median=0.0102,
                             tile_num=2),
             cls.OgcTestCase('S1 IW DES Test',
-                            WmsRequest(data_source=DataSource.SENTINEL1_IW_DES, data_folder=cls.OUTPUT_FOLDER,
+                            WmsRequest(data_collection=DataCollection.SENTINEL1_IW_DES, data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF_d32f, layer='BANDS-S1-IW',
                                        width=img_width, height=img_height, bbox=wgs84_bbox,
                                        time=('2017-10-01', '2017-10-05'),
@@ -183,7 +183,7 @@ class TestOgc(TestSentinelHub):
                             result_len=1, img_min=0.0, img_max=0.07700,
                             img_mean=0.0210801, img_median=0.0132, tile_num=1),
             cls.OgcTestCase('S3 OLCI Test',
-                            WmsRequest(data_source=DataSource.SENTINEL3_OLCI,
+                            WmsRequest(data_collection=DataCollection.SENTINEL3_OLCI,
                                        data_folder=cls.OUTPUT_FOLDER,
                                        image_format=MimeType.TIFF, layer='TRUE-COLOR-S3-OLCI',
                                        width=img_width, height=img_height, bbox=wgs84_bbox_4,
@@ -268,11 +268,11 @@ class TestWebFeatureService(TestSentinelHub):
 
         cls.test_cases = [
             TestCaseContainer('General test',
-                              WebFeatureService(wgs84_bbox, time_interval, data_source=DataSource.SENTINEL2_L1C,
+                              WebFeatureService(wgs84_bbox, time_interval, data_collection=DataCollection.SENTINEL2_L1C,
                                                 maxcc=0.1),
                               result_len=13),
             TestCaseContainer('Latest date only',
-                              WebFeatureService(wgs84_bbox, 'latest', data_source=DataSource.SENTINEL2_L2A),
+                              WebFeatureService(wgs84_bbox, 'latest', data_collection=DataCollection.SENTINEL2_L2A),
                               result_len=1),
         ]
